Refactor: Moves LoginUriActivity to presentation layer of its feature
